// src/enums/Observation.ts
/**
 * Enum for the possible observations while scouting. Multiple can be selected.
 */
enum Observation {
    Tippy = 0,
    DroppingCoral = 1,
    DroppingAlgae = 2,
    DifficultyAligningScore = 3,
    DifficultyAligningIntake = 4,
    Immobilized = 5,
    DisabledPartially = 6,
    DisabledFully = 7
}

export default Observation;